Omg it works... kind of
